+++ integration-tests/Main.hs
@@ -1,5 +1,8 @@
 module Main (main) where
 
+import Test.Tasty
+
+import ResolveCommitRangeTests (resolveCommitRangeTests)
+
 main :: IO ()
-main = do
-  putStrLn "Hello world"
+main = defaultMain $ testGroup "excalibur integration" [resolveCommitRangeTests]