+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to the
 - Fixed per-file checks ignoring the configured `expected-exit` code.
 - Fixed invalid configuration file not causing a non-zero exit code.
 - Added unit tests for many non-IO functions.
+- Added integration tests for many IO functions.
 
 ## 0.1.1 - 2025-10-08